# docs/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ActivateConfirmRequest": {
        "properties": {
          "tenant_id": {
            "maxLength": 6,
            "minLength": 6,
            "title": "Tenant Id",
            "type": "string"
          },
          "user_code": {
            "maxLength": 10,
            "minLength": 8,
            "pattern": "^[A-Z0-9-]{8,10}$",
            "title": "User Code",
            "type": "string"
          }
        },
        "required": [
          "user_code",
          "tenant_id"
        ],
        "title": "ActivateConfirmRequest",
        "type": "object"
      },
      "ActivateConfirmResponse": {
        "properties": {
          "expires_in_seconds": {
            "title": "Expires In Seconds",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "expires_in_seconds"
        ],
        "title": "ActivateConfirmResponse",
        "type": "object"
      },
...
      "Body_token_oauth_token_post": {
        "properties": {
          "client_id": {
            "title": "Client Id",
            "type": "string"
          },
          "device_code": {
            "title": "Device Code",
            "type": "string"
          },
          "grant_type": {
            "title": "Grant Type",
            "type": "string"
          }
        },
        "required": [
          "grant_type",
          "device_code",
          "client_id"
        ],
        "title": "Body_token_oauth_token_post",
        "type": "object"
      },
...
      "DeviceAuthorizationRequest": {
        "description": "Request body for POST /oauth/device_authorization (RFC 8628 §3.1).",
        "properties": {
          "client_id": {
            "const": "agent",
            "title": "Client Id",
            "type": "string"
          },
          "scope": {
            "const": "agent.provision",
            "title": "Scope",
            "type": "string"
          }
        },
        "required": [
          "client_id",
          "scope"
        ],
        "title": "DeviceAuthorizationRequest",
        "type": "object"
      },
      "DeviceAuthorizationResponse": {
        "properties": {
          "device_code": {
            "title": "Device Code",
            "type": "string"
          },
          "expires_in": {
            "exclusiveMinimum": 0.0,
            "maximum": 3600.0,
            "title": "Expires In",
            "type": "integer"
          },
          "interval": {
            "default": 5,
            "maximum": 60.0,
            "minimum": 1.0,
            "title": "Interval",
            "type": "integer"
          },
          "user_code": {
            "title": "User Code",
            "type": "string"
          },
          "verification_uri": {
            "title": "Verification Uri",
            "type": "string"
          },
          "verification_uri_complete": {
            "title": "Verification Uri Complete",
            "type": "string"
          }
        },
        "required": [
          "device_code",
          "user_code",
          "verification_uri",
          "verification_uri_complete",
          "expires_in"
        ],
        "title": "DeviceAuthorizationResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "TokenResponse": {
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "expires_in": {
            "title": "Expires In",
            "type": "integer"
          },
          "refresh_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Refresh Token"
          },
          "token_type": {
            "const": "Bearer",
            "default": "Bearer",
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "expires_in"
        ],
        "title": "TokenResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "CnesData Processing Engine",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/activate/confirm": {
      "post": {
        "operationId": "activate_confirm_activate_confirm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivateConfirmRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivateConfirmResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Activate Confirm",
        "tags": [
          "oauth"
        ]
      }
    },
...
    "/api/v1/jobs/register": {
      "post": {
        "operationId": "register_job_api_v1_jobs_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Body",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Register Job Api V1 Jobs Register Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/v1/jobs/upload-url": {
      "post": {
        "operationId": "mint_upload_url_api_v1_jobs_upload_url_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Body",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Mint Upload Url Api V1 Jobs Upload Url Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mint Upload Url",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/v1/jobs/{job_id}/fail": {
      "post": {
        "operationId": "fail_job_api_v1_jobs__job_id__fail_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Body",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Fail Job Api V1 Jobs  Job Id  Fail Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Fail Job",
        "tags": [
          "jobs"
        ]
      }
    },
...
    "/oauth/device_authorization": {
      "post": {
        "operationId": "device_authorization_oauth_device_authorization_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceAuthorizationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceAuthorizationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Device Authorization",
        "tags": [
          "oauth"
        ]
      }
    },
    "/oauth/token": {
      "post": {
        "operationId": "token_oauth_token_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_token_oauth_token_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Token",
        "tags": [
          "oauth"
        ]
      }
    },
...
  }
}